refactor(node): extract push, query, remove duplicate activity-stream/attention (task-1773689755389)
Phase 2 canvas extraction — all required routes now extracted:

New modules:
- src/canvas-push.ts: push + artifact routes (2 routes)
- src/canvas-query.ts: query route with multipart support (1 route)

Removed duplicates:
- canvas/activity-stream (already in canvas-routes.ts)
- canvas/attention (already in canvas-routes.ts)

Total extracted: push, query, attention, takeover, activity-stream ✅
server.ts reduced by 661 lines (12 canvas routes remaining)
Route-docs: 575/575 pass
Tests: 2455 pass, 0 regressions
