+++ src/canvas-push.ts
@@ -1,0 +1,158 @@
+// SPDX-License-Identifier: Apache-2.0
+// Canvas push + artifact routes — extracted from server.ts
+// Phase 2 canvas route extraction (task-1773689755389)
+
+import type { FastifyInstance } from 'fastify'
+import type { eventBus as eventBusInstance } from './events.js'
+
+interface CanvasPushDeps {
+  eventBus: typeof eventBusInstance
+  queueCanvasPushEvent: (event: Record<string, unknown>) => void
+}
+
+export async function canvasPushRoutes(
+  app: FastifyInstance,
+  deps: CanvasPushDeps,
+): Promise<void> {
+  const { eventBus, queueCanvasPushEvent } = deps
+
+  // POST /canvas/push — agent pushes a visual event to the canvas.
+  // Types: utterance, work_released, handoff, canvas_response, rich
+  app.post('/canvas/push', async (request, reply) => {
+    const body = request.body as Record<string, unknown>
+    const type = typeof body.type === 'string' ? body.type : 'utterance'
+    const agentId = typeof body.agentId === 'string' ? body.agentId.toLowerCase() : 'agent'
+
+    const VALID_PUSH_TYPES = new Set(['utterance', 'work_released', 'handoff', 'canvas_response', 'rich'])
+    if (!VALID_PUSH_TYPES.has(type)) {
+      reply.status(400)
+      return { success: false, message: `type must be one of: ${[...VALID_PUSH_TYPES].join(', ')}` }
+    }
+
+    const now = Date.now()
+    let payload: Record<string, unknown> = { type, agentId, t: now }
+
+    if (type === 'utterance') {
+      const raw = typeof body.text === 'string' ? body.text.trim() : ''
+      const text = raw.slice(0, 60)
+      const ttl = typeof body.ttl === 'number' && body.ttl > 0 ? Math.min(body.ttl, 15_000) : 4_000
+      payload = { ...payload, text, ttl }
+    } else if (type === 'work_released') {
+      const text = typeof body.text === 'string' ? body.text.slice(0, 80) : 'work shipped'
+      const intensity = typeof body.intensity === 'number'
+        ? Math.min(1, Math.max(0.1, body.intensity)) : 0.6
+      const taskTitle = typeof body.taskTitle === 'string' ? body.taskTitle : undefined
+      payload = { ...payload, text, intensity, taskTitle }
+    } else if (type === 'handoff') {
+      const toAgentId = typeof body.toAgentId === 'string' ? body.toAgentId.toLowerCase() : ''
+      if (!toAgentId) {
+        reply.status(400)
+        return { success: false, message: 'handoff requires toAgentId' }
+      }
+      const taskTitle = typeof body.taskTitle === 'string' ? body.taskTitle : undefined
+      const text = typeof body.text === 'string' ? body.text.slice(0, 80) : undefined
+      payload = { ...payload, toAgentId, taskTitle, text }
+    } else if (type === 'rich') {
+      const content = body.content as Record<string, unknown> | undefined
+      if (!content || typeof content !== 'object') {
+        reply.status(400)
+        return { success: false, message: 'rich push requires content object' }
+      }
+      const richContent: Record<string, unknown> = {}
+      if (typeof content.markdown === 'string') richContent.markdown = content.markdown.slice(0, 10_000)
+      if (typeof content.code === 'string') richContent.code = content.code.slice(0, 10_000)
+      if (typeof content.language === 'string') richContent.language = content.language.slice(0, 30)
+      if (typeof content.image === 'string') richContent.image = content.image.slice(0, 2000)
+      if (typeof content.svg === 'string') richContent.svg = content.svg.slice(0, 50_000)
+      if (typeof content.html === 'string') richContent.html = content.html.slice(0, 20_000)
+      if (typeof content.title === 'string') richContent.title = content.title.slice(0, 200)
+
+      const position = typeof body.position === 'object' && body.position
+        ? { x: Number((body.position as any).x) || 0, y: Number((body.position as any).y) || 0 }
+        : undefined
+      const layer = typeof body.layer === 'string' && ['background', 'stage', 'overlay'].includes(body.layer)
+        ? body.layer : 'stage'
+      const ttl = typeof body.ttl === 'number' && body.ttl > 0 ? Math.min(body.ttl, 120_000) : 30_000
+      const size = typeof body.size === 'object' && body.size
+        ? { w: Number((body.size as any).w) || 400, h: Number((body.size as any).h) || 300 }
+        : undefined
+
+      payload = { ...payload, content: richContent, position, layer, ttl, size }
+
+      const RICH_COLORS: Record<string, string> = {
+        link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa', sage: '#34d399',
+        scout: '#fbbf24', echo: '#f472b6', rhythm: '#6ee7b7', spark: '#f97316',
+      }
+      eventBus.emit({
+        id: `cmsg-rich-${now}-${Math.random().toString(36).slice(2, 8)}`,
+        type: 'canvas_message' as const,
+        timestamp: now,
+        data: {
+          type: 'rich',
+          agentId,
+          agentColor: RICH_COLORS[agentId] ?? '#60a5fa',
+          content: richContent,
+          layer,
+        },
+      })
+    } else if (type === 'canvas_response') {
+      const card = body.card as Record<string, unknown> | undefined
+      if (!card || typeof card.type !== 'string') {
+        reply.status(400)
+        return { success: false, message: 'canvas_response requires card with type field' }
+      }
+      const query = typeof body.query === 'string' ? body.query.slice(0, 200) : undefined
+      payload = { ...payload, card, query }
+
+      const RESP_COLORS: Record<string, string> = {
+        link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa', sage: '#34d399',
+        scout: '#f472b6', echo: '#fbbf24', rhythm: '#6ee7b7', spark: '#f97316',
+      }
+      const agentColor = RESP_COLORS[agentId] ?? '#60a5fa'
+      eventBus.emit({
+        id: `cmsg-${now}-${Math.random().toString(36).slice(2, 8)}`,
+        type: 'canvas_message' as const,
+        timestamp: now,
+        data: { ...card, agentId, agentColor, query },
+      })
+    }
+
+    eventBus.emit({ id: `push-${now}-${Math.random().toString(36).slice(2, 6)}`, type: 'canvas_push', timestamp: now, data: payload })
+    queueCanvasPushEvent({ ...payload, _event: 'canvas_push' })
+
+    // Track canvas_first_action activation event (idempotent)
+    const { emitActivationEvent: emitActPush } = await import('./activationEvents.js')
+    emitActPush('canvas_first_action', agentId, { action: 'canvas_push', pushType: type }).catch(() => {})
+
+    return { success: true, type, agentId }
+  })
+
+  // POST /canvas/artifact — emit a proof artifact that drifts through the canvas
+  app.post('/canvas/artifact', async (request, reply) => {
+    const body = request.body as Record<string, unknown>
+    const VALID_TYPES = new Set(['commit', 'pr', 'test', 'run', 'approval'])
+    const type = typeof body.type === 'string' && VALID_TYPES.has(body.type) ? body.type : 'run'
+    const agentId = typeof body.agentId === 'string' ? body.agentId.toLowerCase() : 'agent'
+    const title = typeof body.title === 'string' ? body.title.slice(0, 80) : 'work shipped'
+    const url = typeof body.url === 'string' ? body.url : undefined
+    const taskId = typeof body.taskId === 'string' ? body.taskId : undefined
+    const now = Date.now()
+
+    const AGENT_COLORS: Record<string, string> = {
+      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
+      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
+      rhythm: '#a3e635', swift: '#38bdf8', kotlin: '#f97316',
+    }
+    const agentColor = AGENT_COLORS[agentId] ?? '#94a3b8'
+
+    const payload = { type, agentId, agentColor, title, url, taskId, timestamp: now }
+    eventBus.emit({
+      id: `artifact-${now}-${Math.random().toString(36).slice(2, 6)}`,
+      type: 'canvas_artifact',
+      timestamp: now,
+      data: payload,
+    })
+
+    return { success: true, type, agentId, title }
+  })
+}
