+++ src/server.ts
@@ -11150,151 +11150,8 @@ export async function createServer(): Promise<FastifyInstance> {
     }
   })
 
-  app.get('/canvas/activity-stream', async (request, reply) => {
-    reply.raw.setHeader('Content-Type', 'text/event-stream')
-    reply.raw.setHeader('Cache-Control', 'no-cache')
-    reply.raw.setHeader('Connection', 'keep-alive')
-    reply.raw.setHeader('X-Accel-Buffering', 'no')
-    reply.raw.flushHeaders?.()
+  // canvas/activity-stream + canvas/attention → already in canvas-routes.ts plugin
 
-    let closed = false
-    const subId = `asub-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`
-
-    // Replay backfill — last 20 events with stagger hint for animated replay
-    const backfill = activityRingBuffer.slice(-20)
-    for (let i = 0; i < backfill.length; i++) {
-      if (closed) break
-      try {
-        const entry = { ...backfill[i], _backfill: true, _staggerMs: i * 50 }
-        reply.raw.write(`event: backfill\ndata: ${JSON.stringify(entry)}\n\n`)
-      } catch { break }
-    }
-
-    // Signal backfill complete
-    if (!closed) {
-      try { reply.raw.write(`event: backfill_done\ndata: {}\n\n`) } catch { /* */ }
-    }
-
-    // Register for live events
-    activityStreamSubscribers.set(subId, {
-      closed: false,
-      send: (data: string) => {
-        if (closed) return
-        try { reply.raw.write(`event: activity\ndata: ${data}\n\n`) } catch { closed = true }
-      },
-    })
-
-    request.raw.on('close', () => {
-      closed = true
-      activityStreamSubscribers.delete(subId)
-    })
-
-    return new Promise<void>(() => {})
-  })
-
-  // ── Canvas attention — single highest-priority actionable item ──────────
-  // Returns the one thing that most needs the viewer's attention right now.
-  // Priority: critical notification > high notification > validating task needing review >
-  //           blocked task > medium notification > oldest pending notification
-  // task-1773672750043
-  app.get('/canvas/attention', async (request) => {
-    const query = request.query as Record<string, string>
-    const viewer = typeof query.viewer === 'string' ? query.viewer.trim() : 'human'
-
-    // 1. Check pending notifications (already priority-sorted)
-    const notifModule = await import('./agent-notifications.js')
-    const notifResult = notifModule.getNotifications(getDb(), viewer, { status: 'pending', limit: 1 })
-    const topNotif = notifResult.notifications[0]
-    const notifTotal = notifResult.total
-
-    // 2. Check tasks in validating that viewer could review
-    const validatingTasks = taskManager.listTasks({ status: 'validating' })
-    const reviewable = validatingTasks.find((t: any) =>
-      t.assignee !== viewer && t.reviewers?.includes(viewer)
-    ) ?? validatingTasks[0] // fall back to any validating task
-
-    // 3. Check blocked tasks assigned to viewer
-    const blockedTasks = taskManager.listTasks({ status: 'blocked' })
-    const viewerBlocked = blockedTasks.find((t: any) => t.assignee === viewer)
-
-    // Build attention item — pick highest priority source
-    type AttentionItem = {
-      source: 'notification' | 'review' | 'blocked'
-      priority: 'critical' | 'high' | 'medium' | 'low'
-      title: string
-      detail?: string
-      taskId?: string
-      prUrl?: string
-      agentId?: string
-      actionLabel: string
-      actionType: 'ack' | 'review' | 'unblock'
-      notificationId?: string
-    }
-
-    let item: AttentionItem | null = null
-
-    // Critical/high notifications always win
-    if (topNotif && (topNotif.priority === 'critical' || topNotif.priority === 'high')) {
-      item = {
-        source: 'notification',
-        priority: topNotif.priority,
-        title: topNotif.title,
-        detail: topNotif.body ?? undefined,
-        taskId: topNotif.task_id ?? undefined,
-        agentId: topNotif.source_agent ?? undefined,
-        actionLabel: topNotif.type === 'review' ? 'Review' : 'Acknowledge',
-        actionType: 'ack',
-        notificationId: topNotif.id,
-      }
-    }
-    // Then validating tasks needing review
-    else if (reviewable) {
-      const t = reviewable as any
-      item = {
-        source: 'review',
-        priority: 'high',
-        title: t.title ?? 'Task needs review',
-        detail: `Assigned to ${t.assignee ?? 'unassigned'}`,
-        taskId: t.id,
-        agentId: t.assignee ?? undefined,
-        actionLabel: 'Review',
-        actionType: 'review',
-      }
-    }
-    // Then blocked tasks
-    else if (viewerBlocked) {
-      const t = viewerBlocked as any
-      item = {
-        source: 'blocked',
-        priority: 'medium',
-        title: t.title ?? 'Task is blocked',
-        detail: t.blocked_reason ?? 'Needs attention',
-        taskId: t.id,
-        actionLabel: 'Unblock',
-        actionType: 'unblock',
-      }
-    }
-    // Then any remaining notification
-    else if (topNotif) {
-      item = {
-        source: 'notification',
-        priority: topNotif.priority as 'critical' | 'high' | 'medium' | 'low',
-        title: topNotif.title,
-        detail: topNotif.body ?? undefined,
-        taskId: topNotif.task_id ?? undefined,
-        agentId: topNotif.source_agent ?? undefined,
-        actionLabel: 'Acknowledge',
-        actionType: 'ack',
-        notificationId: topNotif.id,
-      }
-    }
-
-    return { item, pendingCount: notifTotal + validatingTasks.length + blockedTasks.filter((t: any) => t.assignee === viewer).length }
-  })
-
-  // POST /canvas/pulse — agent pushes urgency + optional burst without a full canvas/state update
-  // Lighter-weight than POST /canvas/state; fires canvas_burst if burst=true.
-  // Body: { agentId: string, urgency?: 0–1, burst?: boolean, label?: string }
   app.post('/canvas/pulse', async (request, reply) => {
     const body = request.body as Record<string, unknown>
     const agentId = typeof body.agentId === 'string' ? body.agentId.trim() : ''
@@ -11413,349 +11270,18 @@ export async function createServer(): Promise<FastifyInstance> {
   }
 
   //
-  // Body: { query: string, agentId?: string, sessionId?: string }
-  // Response: { success, card: { type, data, agentId, agentColor } }
-  // Card types: "tasks" | "info" | "revenue" | "onboarding"
-  // SSE event: canvas_message { type, data, agentId, agentColor, query }
-  app.post('/canvas/query', async (request, reply) => {
-    // Support both JSON and multipart/form-data (task-1773673285236)
-    // JSON: { query, attachments: [{ name, type, data (base64) }], sessionId?, agentId? }
-    // Multipart: fields query, sessionId?, agentId? + file parts
-    const contentType = String(request.headers['content-type'] ?? '')
-    const isMultipart = contentType.includes('multipart/form-data')
+  // ── Canvas query route (extracted to src/canvas-query.ts) ─────
+  const { canvasQueryRoutes } = await import("./canvas-query.js")
+  await app.register(canvasQueryRoutes, {
+    eventBus,
+    canvasStateMap,
+    taskManager,
+    chatManager,
+    getCanvasSession,
+    pushCanvasSession,
+    listHosts,
+  } as any)
 
-    let query = ''
-    let sessionIdRaw: string | undefined
-    let agentIdRaw: string | undefined
-    const attachments: Array<{ name: string; type: string; data: string; sizeBytes: number }> = []
-
-    if (isMultipart) {
-      try {
-        const parts = request.parts({ limits: { fileSize: 10 * 1024 * 1024, files: 5 } })
-        for await (const part of parts) {
-          if (part.type === 'field') {
-            const val = typeof part.value === 'string' ? part.value : ''
-            if (part.fieldname === 'query' || part.fieldname === 'text') query = val.trim()
-            else if (part.fieldname === 'sessionId') sessionIdRaw = val
-            else if (part.fieldname === 'agentId') agentIdRaw = val
-          } else if (part.type === 'file') {
-            const chunks: Buffer[] = []
-            for await (const chunk of part.file) chunks.push(chunk)
-            const buf = Buffer.concat(chunks)
-            if (buf.length > 0 && buf.length <= 10 * 1024 * 1024) {
-              attachments.push({
-                name: String(part.filename ?? 'file').slice(0, 255),
-                type: String(part.mimetype ?? 'application/octet-stream'),
-                data: buf.toString('base64'),
-                sizeBytes: buf.length,
-              })
-            }
-          }
-        }
-      } catch (mpErr) {
-        reply.status(400)
-        return { success: false, message: 'Invalid multipart body' }
-      }
-    } else {
-      // JSON path (backward compatible)
-      const body = request.body as Record<string, unknown>
-      query = typeof body.query === 'string' ? body.query.trim() : ''
-      sessionIdRaw = typeof body.sessionId === 'string' ? body.sessionId : undefined
-      agentIdRaw = typeof body.agentId === 'string' ? body.agentId : undefined
-
-      // Extract base64 file attachments from JSON body
-      // Shape: [{ name: string, type: string, data: string (base64) }]
-      const rawAttachments = Array.isArray(body.attachments) ? body.attachments : []
-      for (const att of rawAttachments.slice(0, 5)) { // Max 5 files
-        if (typeof att === 'object' && att && typeof att.name === 'string' && typeof att.data === 'string') {
-          const sizeBytes = Math.ceil((att.data.length * 3) / 4) // base64 → byte estimate
-          if (sizeBytes > 10 * 1024 * 1024) continue // Skip files > 10MB
-          attachments.push({
-            name: String(att.name).slice(0, 255),
-            type: String(att.type || 'application/octet-stream'),
-            data: att.data,
-            sizeBytes,
-          })
-        }
-      }
-    }
-
-    if (!query || query.length > 500) {
-      reply.status(400)
-      return { success: false, message: 'query is required (max 500 chars)' }
-    }
-
-    // Session continuity: client passes sessionId (UUID) so follow-up questions have context
-    const sessionId = sessionIdRaw && sessionIdRaw.length > 0
-      ? sessionIdRaw.trim().slice(0, 64)
-      : null
-    const sessionTurns = sessionId ? getCanvasSession(sessionId) : []
-
-    // Default answering agent is link (builder — knows the codebase + task board)
-    const responderId = agentIdRaw ? agentIdRaw.trim() : 'link'
-    const IDENTITY_COLORS_Q: Record<string, string> = {
-      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
-      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
-    }
-    const agentColor = IDENTITY_COLORS_Q[responderId] ?? '#60a5fa'
-
-    // Gather live context to inject into LLM
-    const allTasksForQuery = taskManager.listTasks({})
-    const activeTasks: Array<{ id: string; title: string; assignee: string; status: string; priority: string }> = []
-    const doingTasks = allTasksForQuery.filter((t: any) => t.status === 'doing').slice(0, 10)
-    const validatingTasks = allTasksForQuery.filter((t: any) => t.status === 'validating').slice(0, 5)
-    for (const t of [...doingTasks, ...validatingTasks] as any[]) {
-      activeTasks.push({ id: t.id, title: t.title ?? '', assignee: t.assignee ?? 'unassigned', status: t.status, priority: t.priority ?? 'P2' })
-    }
-
-    const todoCount = allTasksForQuery.filter((t: any) => t.status === 'todo').length
-    const doingCount = doingTasks.length
-    const validatingCount = validatingTasks.length
-
-    // Build agent orb context
-    const now = Date.now()
-    const STALE_AGENT_MS = 10 * 60 * 1000
-    const activeAgentSummary: string[] = []
-    for (const [agentId, entry] of canvasStateMap) {
-      if (now - entry.updatedAt > STALE_AGENT_MS) continue
-      const payload = entry.payload as Record<string, unknown> ?? {}
-      const state = String((payload as any).presenceState ?? entry.state)
-      const task = (payload as any).activeTask?.title ?? null
-      activeAgentSummary.push(`${agentId}: ${state}${task ? ` — working on "${task.slice(0, 50)}"` : ''}`)
-    }
-
-    // Classify query intent to choose card type
-    const lower = query.toLowerCase()
-    const isTasksQuery = /working on|team doing|team status|happening|active|shipping|tasks|who.?s|what.?s the team/.test(lower)
-    const isRevenueQuery = /revenue|mrr|arr|money|sales|customers|paid|billing/.test(lower)
-    const isOnboardingQuery = /onboard|get started|how do i|where do i start|first step/.test(lower)
-    const isHostsQuery = /show me hosts|host status|server status|machine|node/.test(lower)
-
-    let card: { type: string; data: Record<string, unknown> }
-
-    // Build tasks card from live data (no LLM needed — deterministic)
-    if (isTasksQuery) {
-      const items = activeTasks.slice(0, 5).map(t => ({
-        agentId: t.assignee,
-        agentColor: IDENTITY_COLORS_Q[t.assignee] ?? '#94a3b8',
-        title: t.title,
-        state: t.status,
-      }))
-      const overflow = Math.max(0, activeTasks.length - 5)
-      card = {
-        type: 'tasks',
-        data: { items, overflow, todoCount, doingCount, validatingCount },
-      }
-      // Store summary for session continuity across all card types
-      if (sessionId) {
-        pushCanvasSession(sessionId, 'user', query)
-        pushCanvasSession(sessionId, 'assistant', `${doingCount} tasks in progress, ${validatingCount} validating, ${todoCount} todo.${items.length > 0 ? ` Active: ${items.map(t => t.title.slice(0, 30)).join('; ')}.` : ''}`)
-      }
-    } else if (isRevenueQuery) {
-      // Revenue card — LLM generates honest answer about current state
-      const anthropicKey = process.env.ANTHROPIC_API_KEY
-      let text = 'Revenue tracking not yet wired. Check Stripe directly.'
-      if (anthropicKey) {
-        try {
-          const resp = await fetch('https://api.anthropic.com/v1/messages', {
-            method: 'POST',
-            headers: { 'x-api-key': anthropicKey, 'anthropic-version': '2023-06-01', 'content-type': 'application/json' },
-            body: JSON.stringify({
-              model: 'claude-haiku-4-5',
-              max_tokens: 80,
-              messages: [{ role: 'user', content: `Team Reflectt is a small AI agent team building reflectt.ai (no paid users yet). User asked: "${query}". Honest 1-sentence answer about revenue status. Be direct.` }],
-            }),
-            signal: AbortSignal.timeout(8000),
-          })
-          if (resp.ok) {
-            const d = await resp.json() as { content?: Array<{ text?: string }> }
-            text = d.content?.[0]?.text?.trim() ?? text
-          }
-        } catch { /* use default */ }
-      }
-      card = { type: 'info', data: { text } }
-      if (sessionId) {
-        pushCanvasSession(sessionId, 'user', query)
-        pushCanvasSession(sessionId, 'assistant', text)
-      }
-    } else if (isOnboardingQuery) {
-      card = {
-        type: 'onboarding',
-        data: {
-          step: 1, totalSteps: 3,
-          title: 'Welcome to Reflectt',
-          body: 'Your agents run on reflectt-node. Install it on any machine and your team appears here in the canvas.',
-          ctaLabel: 'Install reflectt-node',
-          ctaAction: 'https://reflectt.ai/docs',
-        },
-      }
-      if (sessionId) {
-        pushCanvasSession(sessionId, 'user', query)
-        pushCanvasSession(sessionId, 'assistant', 'Showing onboarding: install reflectt-node to bring your team to the canvas.')
-      }
-    } else if (isHostsQuery) {
-      const rawHosts = listHosts({})
-      const hosts = rawHosts.map((h: any) => ({
-        id: h.id,
-        name: h.hostname ?? h.id,
-        status: h.status,
-        version: h.version ?? null,
-        agentCount: Array.isArray(h.agents) ? h.agents.length : 0,
-        lastSeen: h.last_seen_at,
-      }))
-      card = { type: 'hosts', data: { hosts } }
-      if (sessionId) {
-        pushCanvasSession(sessionId, 'user', query)
-        const hostSummary = hosts.length > 0
-          ? `${hosts.length} host${hosts.length > 1 ? 's' : ''}: ${hosts.map((h: any) => `${h.name} (${h.status})`).join(', ')}.`
-          : 'No hosts connected yet.'
-        pushCanvasSession(sessionId, 'assistant', hostSummary)
-      }
-    } else {
-      // General query — route to the actual agent via chat.
-      // The agent receives the message in their inbox, processes it through
-      // their real context (OpenClaw session), and can respond via canvas_push.
-      //
-      // This replaces the old standalone LLM call that had no real agent context.
-      // The agents ARE the product — queries go to them, not to a disconnected API key.
-      //
-      // Route: DM to the responder agent on #general (agents subscribe to #general
-      // by default — 'canvas' channel is NOT in DEFAULT_INBOX_SUBSCRIPTIONS, so
-      // messages posted there are never seen by agents).
-      try {
-        const attachmentSummary = attachments.length > 0
-          ? `\n[${attachments.length} file(s) attached: ${attachments.map(a => `${a.name} (${a.type}, ${Math.round(a.sizeBytes / 1024)}KB)`).join(', ')}]`
-          : ''
-        await chatManager.sendMessage({
-          from: 'human',
-          to: responderId,
-          content: `[canvas] @${responderId} ${query}${attachmentSummary}`,
-          channel: 'general',
-          metadata: {
-            source: 'canvas_query',
-            sessionId,
-            responderId,
-            timestamp: Date.now(),
-            reply_via: 'canvas_push', // tells the agent to respond via POST /canvas/push
-            ...(attachments.length > 0 ? { attachments: attachments.map(a => ({ name: a.name, type: a.type, sizeBytes: a.sizeBytes })) } : {}),
-          },
-        })
-      } catch {
-        // Chat delivery failure is non-fatal — still show the thinking card
-      }
-
-      // Return an immediate "thinking" card — the real response will arrive
-      // asynchronously via canvas_push/canvas_message when the agent responds.
-      const text = `Asking ${responderId}…`
-
-      // Store the question in session history
-      if (sessionId) {
-        pushCanvasSession(sessionId, 'user', query)
-      }
-      card = { type: 'info', data: { text, pending: true, responderId } }
-
-      // ── Timeout fallback: if agent doesn't respond within 15s, send a
-      // "no response" card so the UI doesn't hang on "Asking …" forever.
-      let responseReceived = false
-      const listenerId = `canvas-query-timeout-${Date.now()}-${Math.random().toString(36).slice(2, 6)}`
-      eventBus.on(listenerId, (event) => {
-        if (event.type !== 'canvas_message') return
-        const d = event.data as Record<string, unknown> | undefined
-        if (d?.agentId === responderId && d?.isResponse === true) {
-          responseReceived = true
-          eventBus.off(listenerId)
-        }
-      })
-
-      setTimeout(() => {
-        eventBus.off(listenerId)
-        if (responseReceived) return
-        // Emit a timeout fallback card
-        eventBus.emit({
-          id: `cmsg-timeout-${Date.now()}`,
-          type: 'canvas_message' as const,
-          timestamp: Date.now(),
-          data: {
-            type: 'info',
-            data: { text: `${responderId} is busy right now. Try again in a moment, or ask a different agent.`, pending: false },
-            agentId: responderId,
-            agentColor,
-            isResponse: true,
-            isTimeout: true,
-          },
-        })
-        if (sessionId) {
-          pushCanvasSession(sessionId, 'assistant', `(${responderId} did not respond within 15s)`)
-        }
-      }, 15_000)
-    }
-
-    // Emit canvas_message on event bus — pulse stream forwards it to all subscribers
-    eventBus.emit({
-      id: `cmsg-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`,
-      type: 'canvas_message' as const,
-      timestamp: Date.now(),
-      data: {
-        ...card,
-        agentId: responderId,
-        agentColor,
-        query,
-        ...(attachments.length > 0 ? { attachments: attachments.map(a => ({ name: a.name, type: a.type, sizeBytes: a.sizeBytes })) } : {}),
-      },
-    })
-
-    return { success: true, card: { ...card, agentId: responderId, agentColor, ...(attachments.length > 0 ? { attachmentCount: attachments.length } : {}) } }
-  })
-
-  // ── Canvas query response bridge ───────────────────────────────────────────
-  // When an agent responds to a [canvas] query (via chat), convert their response
-  // into a canvas_message event so the browser canvas can display it.
-  // This bridges: agent chat response → canvas card.
-  eventBus.on('canvas-query-response-bridge', (event) => {
-    if (event.type !== 'message_posted') return
-    const data = event.data as Record<string, unknown>
-    const content = String(data.content ?? '')
-    const from = String(data.from ?? '')
-    const channel = String(data.channel ?? '')
-
-    // Only bridge messages from agents (not from 'human' or 'system')
-    if (from === 'human' || from === 'system' || from === 'github') return
-
-    // Detect canvas responses: messages that start with [canvas-response] or
-    // are on the canvas channel from an agent, or mention [canvas] in reply
-    const isCanvasResponse = content.startsWith('[canvas-response]')
-      || content.startsWith('[canvas]')
-      || (channel === 'canvas' && from !== 'human')
-    if (!isCanvasResponse) return
-
-    // Strip the [canvas-response] / [canvas] prefix
-    const cleanContent = content
-      .replace(/^\[canvas-response\]\s*/i, '')
-      .replace(/^\[canvas\]\s*/i, '')
-      .trim()
-    if (!cleanContent) return
-
-    const IDENTITY_COLORS_BRIDGE: Record<string, string> = {
-      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
-      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
-      rhythm: '#a3e635', swift: '#38bdf8',
-    }
-    const agentColor = IDENTITY_COLORS_BRIDGE[from] ?? '#94a3b8'
-
-    // Emit as canvas_message — browser pulse stream picks it up
-    eventBus.emit({
-      id: `cmsg-response-${Date.now()}-${Math.random().toString(36).slice(2, 6)}`,
-      type: 'canvas_message' as const,
-      timestamp: Date.now(),
-      data: {
-        type: 'info',
-        data: { text: cleanContent },
-        agentId: from,
-        agentColor,
-        isResponse: true,
-      },
-    })
-  })
 
   // POST /canvas/push — agent self-initiates a canvas event without a human query.
   // Agents call this to surface their own work: utterances that float from their orb,
@@ -11763,181 +11289,12 @@ export async function createServer(): Promise<FastifyInstance> {
   // All events emit on the pulse SSE stream as canvas_push for the browser to render.
   //
   // pixel spec: design/canvas-as-ours.html
-  //
-  // Body:
-  //   type: 'utterance' | 'work_released' | 'handoff'
-  //   agentId: string
-  //   text?: string              (utterance: max 60 chars; work_released: optional label)
-  //   ttl?: number               (utterance: default 4000ms)
-  //   intensity?: number 0–1     (work_released: effort weight; default 0.6)
-  //   toAgentId?: string         (handoff: receiving agent)
-  //   taskTitle?: string         (handoff + work_released: what moved)
-  app.post('/canvas/push', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const type = typeof body.type === 'string' ? body.type : 'utterance'
-    const agentId = typeof body.agentId === 'string' ? body.agentId.toLowerCase() : 'agent'
-
-    const VALID_PUSH_TYPES = new Set(['utterance', 'work_released', 'handoff', 'canvas_response', 'rich'])
-    if (!VALID_PUSH_TYPES.has(type)) {
-      reply.status(400)
-      return { success: false, message: `type must be one of: ${[...VALID_PUSH_TYPES].join(', ')}` }
-    }
-
-    const now = Date.now()
-
-    let payload: Record<string, unknown> = { type, agentId, t: now }
-
-    if (type === 'utterance') {
-      const raw = typeof body.text === 'string' ? body.text.trim() : ''
-      const text = raw.slice(0, 60)  // max 60 chars per spec
-      const ttl = typeof body.ttl === 'number' && body.ttl > 0 ? Math.min(body.ttl, 15_000) : 4_000
-      payload = { ...payload, text, ttl }
-    } else if (type === 'work_released') {
-      const text = typeof body.text === 'string' ? body.text.slice(0, 80) : 'work shipped'
-      const intensity = typeof body.intensity === 'number'
-        ? Math.min(1, Math.max(0.1, body.intensity)) : 0.6
-      const taskTitle = typeof body.taskTitle === 'string' ? body.taskTitle : undefined
-      payload = { ...payload, text, intensity, taskTitle }
-    } else if (type === 'handoff') {
-      const toAgentId = typeof body.toAgentId === 'string' ? body.toAgentId.toLowerCase() : ''
-      if (!toAgentId) {
-        reply.status(400)
-        return { success: false, message: 'handoff requires toAgentId' }
-      }
-      const taskTitle = typeof body.taskTitle === 'string' ? body.taskTitle : undefined
-      const text = typeof body.text === 'string' ? body.text.slice(0, 80) : undefined
-      payload = { ...payload, toAgentId, taskTitle, text }
-    } else if (type === 'rich') {
-      // Rich content — agents push arbitrary visual content to the canvas.
-      // This is the "agents control every pixel" path. Content can be markdown,
-      // code blocks, images, SVG, or raw HTML. The canvas renderer interprets it.
-      // task-1773672750043
-      const content = body.content as Record<string, unknown> | undefined
-      if (!content || typeof content !== 'object') {
-        reply.status(400)
-        return { success: false, message: 'rich push requires content object' }
-      }
-      // Sanitize content fields
-      const richContent: Record<string, unknown> = {}
-      if (typeof content.markdown === 'string') richContent.markdown = content.markdown.slice(0, 10_000)
-      if (typeof content.code === 'string') richContent.code = content.code.slice(0, 10_000)
-      if (typeof content.language === 'string') richContent.language = content.language.slice(0, 30)
-      if (typeof content.image === 'string') richContent.image = content.image.slice(0, 2000) // URL only
-      if (typeof content.svg === 'string') richContent.svg = content.svg.slice(0, 50_000)
-      if (typeof content.html === 'string') richContent.html = content.html.slice(0, 20_000)
-      if (typeof content.title === 'string') richContent.title = content.title.slice(0, 200)
-
-      // Position and display hints
-      const position = typeof body.position === 'object' && body.position
-        ? { x: Number((body.position as any).x) || 0, y: Number((body.position as any).y) || 0 }
-        : undefined
-      const layer = typeof body.layer === 'string' && ['background', 'stage', 'overlay'].includes(body.layer)
-        ? body.layer : 'stage'
-      const ttl = typeof body.ttl === 'number' && body.ttl > 0 ? Math.min(body.ttl, 120_000) : 30_000
-      const size = typeof body.size === 'object' && body.size
-        ? { w: Number((body.size as any).w) || 400, h: Number((body.size as any).h) || 300 }
-        : undefined
-
-      payload = { ...payload, content: richContent, position, layer, ttl, size }
-
-      // Also emit as canvas_message for activity feed visibility
-      const RICH_COLORS: Record<string, string> = {
-        link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa', sage: '#34d399',
-        scout: '#fbbf24', echo: '#f472b6', rhythm: '#6ee7b7', spark: '#f97316',
-      }
-      eventBus.emit({
-        id: `cmsg-rich-${now}-${Math.random().toString(36).slice(2, 8)}`,
-        type: 'canvas_message' as const,
-        timestamp: now,
-        data: {
-          type: 'rich',
-          agentId,
-          agentColor: RICH_COLORS[agentId] ?? '#60a5fa',
-          content: richContent,
-          layer,
-        },
-      })
-    } else if (type === 'canvas_response') {
-      // Agent responds to a canvas query with a structured card.
-      // This is how agents answer questions typed on the canvas —
-      // the query arrives via chat, agent processes it, responds here.
-      const card = body.card as Record<string, unknown> | undefined
-      if (!card || typeof card.type !== 'string') {
-        reply.status(400)
-        return { success: false, message: 'canvas_response requires card with type field' }
-      }
-      const query = typeof body.query === 'string' ? body.query.slice(0, 200) : undefined
-      payload = { ...payload, card, query }
-
-      // Also emit as canvas_message so living-canvas renders it as a response card
-      // (same event type as the old synchronous canvas/query response)
-      const RESP_COLORS: Record<string, string> = {
-        link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa', sage: '#34d399',
-        scout: '#f472b6', echo: '#fbbf24', rhythm: '#6ee7b7', spark: '#f97316',
-      }
-      const agentColor = RESP_COLORS[agentId] ?? '#60a5fa'
-      eventBus.emit({
-        id: `cmsg-${now}-${Math.random().toString(36).slice(2, 8)}`,
-        type: 'canvas_message' as const,
-        timestamp: now,
-        data: { ...card, agentId, agentColor, query },
-      })
-    }
-
-    // Emit on eventBus — forwarded immediately on pulse SSE stream (local SSE subscribers)
-    eventBus.emit({ id: `push-${now}-${Math.random().toString(36).slice(2, 6)}`, type: 'canvas_push', timestamp: now, data: payload })
-
-    // Queue for cloud relay — reaches browsers on app.reflectt.ai via syncCanvas push_events[]
-    // task-1773690756100
-    queueCanvasPushEvent({ ...payload, _event: 'canvas_push' })
-
-    // Track canvas_first_action activation event (idempotent — fires once per agentId)
-    // task-1773692063045-f3ggtwnbr
-    const { emitActivationEvent: emitActPush } = await import('./activationEvents.js')
-    emitActPush('canvas_first_action', agentId, { action: 'canvas_push', pushType: type }).catch(() => {})
-
-    return { success: true, type, agentId }
-  })
-
-  // POST /canvas/artifact — emit a proof artifact that drifts through the canvas.
-  // Fires automatically on task completion and PR merge (see hooks below).
-  // Agents can also call this directly to surface any work artifact.
-  //
-  // spec: design/interface-os-v0-artifact-stream.html
-  //
-  // Body:
-  //   type: 'commit' | 'pr' | 'test' | 'run' | 'approval'
-  //   agentId: string          (sender agent)
-  //   title: string            (short label, max 80 chars)
-  //   url?: string             (link to artifact)
-  //   taskId?: string          (related task, for context)
-  app.post('/canvas/artifact', async (request, reply) => {
-    const body = request.body as Record<string, unknown>
-    const VALID_TYPES = new Set(['commit', 'pr', 'test', 'run', 'approval'])
-    const type = typeof body.type === 'string' && VALID_TYPES.has(body.type) ? body.type : 'run'
-    const agentId = typeof body.agentId === 'string' ? body.agentId.toLowerCase() : 'agent'
-    const title = typeof body.title === 'string' ? body.title.slice(0, 80) : 'work shipped'
-    const url = typeof body.url === 'string' ? body.url : undefined
-    const taskId = typeof body.taskId === 'string' ? body.taskId : undefined
-    const now = Date.now()
-
-    const AGENT_COLORS: Record<string, string> = {
-      link: '#60a5fa', kai: '#fb923c', pixel: '#a78bfa',
-      sage: '#34d399', scout: '#fbbf24', echo: '#f472b6',
-      rhythm: '#a3e635', swift: '#38bdf8', kotlin: '#f97316',
-    }
-    const agentColor = AGENT_COLORS[agentId] ?? '#94a3b8'
-
-    const payload = { type, agentId, agentColor, title, url, taskId, timestamp: now }
-    eventBus.emit({
-      id: `artifact-${now}-${Math.random().toString(36).slice(2, 6)}`,
-      type: 'canvas_artifact',
-      timestamp: now,
-      data: payload,
-    })
-
-    return { success: true, type, agentId, title }
-  })
+  // ── Canvas push + artifact routes (extracted to src/canvas-push.ts) ─────
+  const { canvasPushRoutes } = await import("./canvas-push.js")
+  await app.register(canvasPushRoutes, {
+    eventBus,
+    queueCanvasPushEvent,
+  } as any)
 
   // GET /canvas/pulse — SSE stream emitting a heartbeat tick every 2s with live intensity values
   // Drives smooth canvas animation without polling. Each tick includes per-agent orb data + team mood.
